requirements.txt
django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
yt-dlp==2025.10.22
uvicorn==0.35.0